requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
plotly
python-telegram-bot
garminconnect
requests
python-dotenv
jinja2
